chore(refactor): create cmd folders for dataproxy & dataproxyupdater
