+++ cmd/dataproxy/dataproxy/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/MurmurationsNetwork/MurmurationsServices/services/dataproxy/pkg/dataproxy"
+
+func main() {
+	dataproxy.StartApplication()
+}
