+++ cmd/dataproxyupdater/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/MurmurationsNetwork/MurmurationsServices/services/dataproxyupdater/pkg/dataproxyupdater"
+
+func main() {
+	dataproxyupdater.Run()
+}
