+++ docs/seed.md
@@ -32,11 +32,11 @@ Example request body:
    ```
    ./seeder "https://docs.google.com/uc?export=download&id=[EXCEL_ID]" "karte_von_morgen-v1.0.0" 2 2
    ```
-   
+
 ## Seed data in local
 1. List the pods and find the data-proxy-app. `kubectl get pods`
 2. Connect to the data-proxy-app pod. `kubectl exec -it [POD_NAME] -- sh`
-3. Change directory to seeder folder. `cd services/dataproxy/cmd/seeder`
+3. Change directory to seeder folder. `cd cmd/dataproxy/seeder`
 4. Execute the command with the format `go run main.go "EXCEL_URL" "SCHEMA_NAME" "FROM" "TO"`
    The following is an example:
    ```
