+++ scripts/determine_exclusions.sh
@@ -14,11 +14,11 @@ declare -A SERVICE_PATHS=(
     ["library"]="go.mod pkg/ cmd/library/ services/library/"
     ["geoip"]="go.mod pkg/ cmd/geoip/ services/geoip/"
     ["validation"]="go.mod pkg/ cmd/validation/ services/validation/"
-    ["dataproxy"]="go.mod pkg/ services/dataproxy/"
+    ["dataproxy"]="go.mod pkg/ cmd/dataproxy/ services/dataproxy/"
     ["nodecleaner"]="go.mod pkg/ cmd/nodecleaner/ services/nodecleaner/"
     ["revalidatenode"]="go.mod pkg/ cmd/revalidatenode/ services/revalidatenode/"
     ["schemaparser"]="go.mod pkg/ cmd/schemaparser/ services/schemaparser/"
-    ["dataproxyupdater"]="go.mod pkg/ services/dataproxyupdater/"
+    ["dataproxyupdater"]="go.mod pkg/ cmd/dataproxyupdater/ services/dataproxyupdater/"
     ["dataproxyrefresher"]="go.mod pkg/ cmd/dataproxyrefresher/ services/dataproxyrefresher/"
 )
