--- services/dataproxy/cmd/dataproxy/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/MurmurationsNetwork/MurmurationsServices/services/dataproxy/internal/app"
-
-func main() {
-	app.StartApplication()
-}
