+++ services/dataproxy/pkg/dataproxy/application.go
@@ -1,4 +1,4 @@
-package app
+package dataproxy
 
 import (
 	"fmt"
