+++ services/dataproxy/pkg/dataproxy/cleanup.go
@@ -1,4 +1,4 @@
-package app
+package dataproxy
 
 func cleanup() {
 }
