+++ services/dataproxy/pkg/dataproxy/graceful_shutdown.go
@@ -1,4 +1,4 @@
-package app
+package dataproxy
 
 import (
 	"context"
