+++ services/dataproxy/pkg/dataproxy/middleware.go
@@ -1,4 +1,4 @@
-package app
+package dataproxy
 
 import (
 	"time"
