--- services/dataproxy/reflex.conf
@@ -1,1 +1,0 @@
--r '\.go' -R '_test\.go$' -s go run ./services/dataproxy/cmd/dataproxy
